Add local subject heading logic
