+++ .rubocop.yml
@@ -20,7 +20,7 @@ Metrics/AbcSize:
         - 'app/services/oai_processing_service.rb'
         - 'app/services/oai_processing_single_service.rb'
         - 'lib/chicago_citation_formatter.rb'
-        - 'lib/traject/extract_subject_display.rb'
+        - 'lib/traject/extract_subject.rb'
         - 'app/helpers/citation_modal_helper.rb'
 
 Metrics/BlockLength:
@@ -65,6 +65,7 @@ Metrics/ModuleLength:
     Exclude:
         - 'app/models/concerns/blacklight/solr/document/marc_export.rb'
         - 'lib/traject/extraction_tools.rb'
+        - 'lib/traject/extract_subject.rb'
         - 'config/prepends/custom_citation_logic.rb'
         - 'app/models/concerns/statusable.rb'
         - 'app/helpers/catalog_helper.rb'
